Add web_edit_button_conditional_disable
A new module that allows other modules that depend on it
to define a boolean field `should_disable_form_edit` to disable
the editing button on form views.

This module is based on OCA's module `web_access_rule_buttons`,
but instead of disabling editing based on record access,
allows disabling to be based on anything that can be represented
as a boolean field on the record.

Original module web_access_rule_buttons can be found in OCA's
web repository here: https://github.com/OCA/web/tree/14.0/web_access_rule_buttons